change the type of features to double and type of labels to int of new datasets.

+++ utils/load_dataset.py
@@ -498,80 +498,80 @@ def load_covtype():
 def load_9_Tumors():
     data = sio.loadmat('UCI Data/Gems/9_Tumors.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_11_Tumors():
     data = sio.loadmat('UCI Data/Gems/11_Tumors.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_14_Tumors():
     data = sio.loadmat('UCI Data/Gems/14_Tumors.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_Brain_Tumor1():
     data = sio.loadmat('UCI Data/Gems/Brain_Tumor1.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_DLBCL():
     data = sio.loadmat('UCI Data/Gems/DLBCL.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_Leukemia1():
     data = sio.loadmat('UCI Data/Gems/Leukemia1.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_Leukemia2():
     data = sio.loadmat('UCI Data/Gems/Leukemia2.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_Lung_Cancer():
     data = sio.loadmat('UCI Data/Gems/Lung_Cancer.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_Prostate_Tumor():
     data = sio.loadmat('UCI Data/Gems/Prostate_Tumor.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 
 def load_SRBCT():
     data = sio.loadmat('UCI Data/Gems/SRBCT.mat')
     fea = data['data']
-    labels = fea[:, 0]
-    fea = fea[:, 1:]
+    labels = fea[:, 0].astype(int)
+    fea = fea[:, 1:].astype(np.float64)
     return fea, labels
 
 def Test():
